Implement structs
Add an additional AST_STRUCT_TYPE that allows a sequence of fields of
various types to be aggregated into a new type. Also add a AST_TYPE_UNIT
that allows the programmer to define aliases for types. Implement the
member access operator such that variables of struct type may have their
member fields' values changed.

Add a list of unimplemented features to the README.