Implement import statement
For each import statement in a file, parse the imported file and perform
codegeneration for that file's declarations, which basically makes the
types and function prototypes that the file declares visible. Make sure
that each file is only imported once. This allows code to refer to types
and functions that are declared in a different file. After the call to
lowc, the resulting files must be linked in order to resolve all
references.

Add an example project that compiles two .low files into a library that
is then used from a third .low file. This makes use of the llvm-link
tool.

Add stdio.low which contains a translation of stdio.h into Low.