Implement lexical consts
Implement lexical constants like they appear in the Go language. That
is, constants consist of an expression but no type, and that are
assigned a type only when they are referenced somewhere in the code.
This allows number literals such as `42` or `3.141` to be typeless and
be converted to the appropriate type (e.g. `int8` or `float32`) when
they are used. To achieve this, quite a few changes to the code
generation procedure are necessary. Especially the determination of
types needs to occur in a separate pass.