Add float support and improve cast handling
Add support for floating point types, namely 16, 32, 64, and 128 bits in
width. Also add a `bool` type that saturates at 0 and 1 and only
supports boolean operations. This makes things pretty typesafe and
avoids a few pitfalls of C. Adapt the code generation for binary
expressions such that they handle integers, floats, and bools
separately. Most operators are not defined on all types, which makes
this even trickier. The current implementation does not allow for
implicit casts to bool, which might be a good or a bad thing.

While we're at it, implement the `<< >> % & ^ | && ||` binary operators.