Implement sizeof operator
Add code generation for the sizeof operator. The grammar rules are
fairly close to those of C, with an exception. Since C's grammar is not
context-free, it is able to distinguish between type names and
identifiers. In C, you either use `sizeof(<type>)` or `sizeof <expr>`.
This is not possible in context-free Low. Therefore we drop the
parentheses and add the hashtag `#` as a marker for the type version of
the operator: `sizeof #<type>` or `sizeof <expr>`. Note that sizeof acts
as a unary operator, with all the corresponding precedence rules. That
causes the expression `sizeof #(int8)0` to fail, since the cast operator
has lower precedence than the sizeof operator. Maybe we should introduce
parentheses in the grammar to enforce `sizeof(#<type>|<expr>)` for
better readability?