Add third kind of tests that need to execute
Until now we have two kinds of tests: Those that should compile, and
those that should fail to compile. Add a third kind of test that should
compile and then return exit code 0 when executed. This allows for
checks to be implementd in the Low language.

Add a test that checks whether new returns zeroed memory to test this
new facility.