Configure continuous integration testing
Add a `.travis.yml` file as per issue #9 to enable continuous
integration builds and tests. This will allow us to catch errors early
on, and also verify that the project builds with both gcc and clang.

Also add an introductory paragraph to the README file, together with the
travis badge.