Implement function pointers
Extend the grammar, code generation, and the type system to allow for
function pointers to be called. As in C, simply typing a function's name
will yield a pointer to it.

Fix the sizeof() operation such that it honors the type suggested by the
context in which it is used. Note that this is an implicit downcast.
Maybe we can verify that the size always fits in the suggested int
statically?

Also add a test case that check whether the argument to return receives
a type hint based on the return type of the function.