Implement separate test script
Change the way tests are executed by adding a script that will iterate
over each *.low file in the test/ directory and perform tests on it. The
test is configured from within the *.low file, by placing the markers
"+compile", "-compile", "+execute", or "-execute" somewhere in the file.
Tests are compiled and optionally run, with exit codes interpreted
according to the +/- prefix. This makes the folder structure in test/
redundant, but it can be kept for clarity if desired.