Rename test makefile target
Due to a collision with CMake's internal testing facilities (CTest), we
cannot have a target called "test" without CMake complaining. This could
be disabled via a policy instruction, but that instruction would fail on
older CMake versions that do not have that policy in place yet. Thus the
easiest way to cope with this should be to rename the target "test" into
"tests" in the CMakeLists.txt and adjust the `.travis.sh` accordingly.