Add the boolean constants 'true' and 'false'
Tweak the code generation step to allow for the identifiers `true` and
`false` to be used to refer to a constant 0 or 1 of the `bool` type.
Using true and false in code is superior to 1 and 0, since the former
are typed and may be used to deduce a variable's type.