Implement interfaces for struct members
Add an interface type that can hold a pointer to an arbitrary struct in
combination with a pointer to a table holding information about the
members of said struct. This allows a function to ask for "something
that has an x and y member" and be given arbitrary structs that fulfill
this criterion.