Consolidate loops and allow empty exprs
Generalize the for and while loops such that `for [cond] {}` and
`for [init];[cond];[step] {}` are allowed. To do this, fix an issue in
the grammar which prevented the initial, condition, and step expressions
from being allowed to be empty. This requires an additional fix in the
code generation part, such that for loops that lack a condition would
unconditionally branch into the loop body, and would skip code
generation for the initial and step expressions if none are present.

It should now be possible to drop the separate AST_WHILE_STMT, since its
functionality can be achieved via AST_FOR_STMT.