Implement interface functions
Add a first implementation of interface functions. Interfaces may now
declare functions that will be included in their lookup table. The whole
implementation is a bit hacky, but is a fairly low-level and general
approach to the concept. Currently, the list of functions that
correspond to a specific function in an interface needs to be listed
explicitly for each type. While this possibility shall remain, a more
convenient approach is needed to facilitate this.