Refactor codegen
Move the codegen code for each of the expression AST nodes into separate
files rather than having everything in a file of 2000 lines. Also rename
`determine_type` to `prepare_expr`, suggesting that identifiers and
function calls should look for their targets during this step and then
store the results in the AST node. This should allow us to remove the
redundant name lookup in many of the codegen parts.