Add link option
Implement command line option parsing and add an option to link all
input files into a single object file that is written to disk. Generally
speaking, lowc compiles every source file independently and emits the
compiled LLVM IR to disk. However, to simplify use, the compiler may be
asked for a linked version of the input files provided with the '-o'
command line option. Future versions of the compiler shall provide
additional options that allow for libraries and executables to be
emitted directly.