Implement array slicing, change slice structure
Extend the slicing operator `[a:b]` to support creating a slice of an
array. This allows the backing storage of a slice to reside on the
stack, which is useful for embedded systems. To simplify the code a bit,
also adjust the structure of a slice, rather than having `[0 x T]*` as
the type of the backing storage, use `T*`.