Remove trailing 0 byte from string length
Always have a trailing 0 byte at the end of a string, but do not account
for it in the length of the slice. This makes working with strings much
more natural and still maintains compatibility with C code. All code
modifying strings must ensure that the s[len(s)] byte is set to 0 and
that cap(s) >= len(s)+1.

Also add the scaffolding for a unicode library, which remains empty for
now.